add yum & alpine based linux support (#91)
* add yum based linux support

* add centos 7 tests

* add alpine support

* try doing a different approach

+++ src/runaqa.ts
@@ -47,9 +47,15 @@ export async function runaqaTest(
   let dependents = await tc.downloadTool(
     'https://ci.adoptopenjdk.net/view/all/job/test.getDependency/lastSuccessfulBuild/artifact//*zip*/dependents.zip'
   )
+
+  let sevenzexe = '7z'
+  if (fs.existsSync('/usr/bin/yum')) {
+    sevenzexe = '7za'
+  }
+
   // Test.dependency only has one level of archive directory, none of actions toolkit support mv files by regex. Using 7zip discards the directory directly
   await exec.exec(
-    `7z e ${dependents} -o${process.env.GITHUB_WORKSPACE}/aqa-tests/TKG/lib`
+    `${sevenzexe} e ${dependents} -o${process.env.GITHUB_WORKSPACE}/aqa-tests/TKG/lib`
   )
 
   if (buildList.includes('system')) {
@@ -96,7 +102,11 @@ export async function runaqaTest(
 }
 
 function getTestJdkHome(version: string, jdksource: string): string {
+  // Try JAVA_HOME first and then fall back to GITHUB actions default location
   let javaHome = process.env[`JAVA_HOME_${version}_X64`] as string
+  if (javaHome === undefined) {
+    javaHome = process.env['JAVA_HOME'] as string
+  }
   if (jdksource === 'install-jdk') {
     // work with AdoptOpenJDK/install-sdk
     if (`JDK_${version}` in process.env) {
@@ -108,6 +118,9 @@ function getTestJdkHome(version: string, jdksource: string): string {
   // Remove spaces in Windows path and replace with a short name path, e.g. 'C:/Program Files/***' ->C:/Progra~1/***
   if (IS_WINDOWS && jdksource === 'github-hosted') {
     javaHome = javaHome.replace(/Program Files/g, 'Progra~1')
+  }
+  if (javaHome === undefined) {
+    core.error('JDK could not be found')
   }
   return javaHome
 }
@@ -150,8 +163,35 @@ async function installDependencyAndSetup(): Promise<void> {
     await exec.exec('sudo sysctl -w kern.sysv.shmall=655360')
     await exec.exec('sudo sysctl -w kern.sysv.shmmax=125839605760')
   } else {
-    await exec.exec('sudo apt-get update')
-    await exec.exec('sudo apt-get install ant-contrib -y')
+    if (fs.existsSync('/usr/bin/apt-get')) {
+      // Debian Based
+      await exec.exec('sudo apt-get update')
+      await exec.exec('sudo apt-get install ant-contrib -y')
+    } else if (fs.existsSync('/usr/bin/yum')) {
+      // RPM Based
+      await exec.exec('sudo yum update -y')
+      await exec.exec('sudo yum install p7zip -y')
+      const antContribFile = await tc.downloadTool(
+        `https://sourceforge.net/projects/ant-contrib/files/ant-contrib/ant-contrib-1.0b2/ant-contrib-1.0b2-bin.zip/download`
+      )
+      await tc.extractZip(`${antContribFile}`, `${tempDirectory}`)
+      await io.cp(
+        `${tempDirectory}/ant-contrib/lib/ant-contrib.jar`,
+        `${process.env.ANT_HOME}\\lib`
+      )
+    } else if (fs.existsSync('/sbin/apk')) {
+      // Alpine Based
+      await exec.exec('apk update')
+      await exec.exec('apk add p7zip')
+      const antContribFile = await tc.downloadTool(
+        `https://sourceforge.net/projects/ant-contrib/files/ant-contrib/ant-contrib-1.0b2/ant-contrib-1.0b2-bin.zip/download`
+      )
+      await tc.extractZip(`${antContribFile}`, `${tempDirectory}`)
+      await io.cp(
+        `${tempDirectory}/ant-contrib/lib/ant-contrib.jar`,
+        `${process.env.ANT_HOME}\\lib`
+      )
+    }
     //environment
     if ('RUNNER_USER' in process.env) {
       process.env['LOGNAME'] = process.env['RUNNER_USER']
@@ -161,8 +201,10 @@ async function installDependencyAndSetup(): Promise<void> {
       )
     }
 
-    //disable apport
-    await exec.exec('sudo service apport stop')
+    if (fs.existsSync('/usr/bin/apt-get')) {
+      //disable apport
+      await exec.exec('sudo service apport stop')
+    }
   }
 }
 
@@ -187,7 +229,11 @@ async function getOpenjdkTestRepo(openjdktestRepo: string): Promise<void> {
   process.chdir('aqa-tests')
 }
 
-async function runGetSh(tkgRepo: string, openj9Repo: string, vendorTestParams: string): Promise<void> {
+async function runGetSh(
+  tkgRepo: string,
+  openj9Repo: string,
+  vendorTestParams: string
+): Promise<void> {
   let parameters = ''
   if (tkgRepo !== 'TKG:master') {
     const repoBranch = parseRepoBranch(tkgRepo)
